+++ golem-worker-service-base/src/parser/expr/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod expr_parser;
 mod flags;
 mod let_statement;
+mod pattern_match;
 mod record;
 mod selection;
 mod sequence;
